chore: deduplicate mappers and model
